partner_bot/requirements.txt
pyttsx3
pygame
nest_asyncio
requests
gtts
google-genai
google-api-python-client
python-telegram-bot
playsound==1.2.2
$env:GEMINI_API_KEY="あなたのAPIキー"